dom/workers/test/test_throwingOnerror.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads
-->
<head>
  <title>Test for DOM Worker Threads Recursion</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  var worker = new Worker("throwingOnerror_worker.js");

  var errors = ["foo", "bar"];

  worker.onerror = function(event) {
    event.preventDefault();
    var found = false;
    for (var index in errors) {
      if (event.message == "uncaught exception: " + errors[index]) {
        errors.splice(index, 1);
        found = true;
        break;
      }
    }
    is(found, true, "Unexpected error!");
  };

  worker.onmessage = function(event) {
    is(errors.length, 0, "Didn't see expected errors!");
    SimpleTest.finish();
  };

  for (var i = 0; i < 2; i++) {
    worker.postMessage("");
  }

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
